+++ app/containers/App/index.js
@@ -12,11 +12,21 @@
  */
 
 import React from 'react';
-export default class App extends React.Component { // eslint-disable-line react/prefer-stateless-function
+import { connect } from 'react-redux';
+import { createStructuredSelector } from 'reselect';
+import { loadData } from './actions';
+
+import { makeSelectData } from './selectors';
+export class App extends React.Component { // eslint-disable-line react/prefer-stateless-function
 
   static propTypes = {
     children: React.PropTypes.node,
   };
+  componentWillMount() {
+    this.props.getAllData();
+    // this.props.getXrmData();
+  }
+  
   render() {
     return (
       <div>
@@ -28,5 +38,22 @@ export default class App extends React.Component { // eslint-disable-line react/
 
 App.propTypes = {
   children: React.PropTypes.node,
+  dispatch: React.PropTypes.func.isRequired,
+  getAllData: React.PropTypes.func,
 };
 
+const mapStateToProps = createStructuredSelector({
+  data: makeSelectData(),
+
+});
+
+function mapDispatchToProps(dispatch) {
+  return {
+    dispatch,
+    getAllData: () => {
+      dispatch(loadData());
+    },
+  };
+}
+
+export default connect(mapStateToProps, mapDispatchToProps)(App);
