+++ app/containers/EditQuotePage/constants.js
@@ -12,3 +12,5 @@
 // export const DELETE_LINE = 'app/EditQuote/DELETE_LINE';
 // export const LOAD_XRM_DATA = 'app/EditQuote/LOAD_XRM_DATA';
 // export const LOAD_XRM_DATA_SUCCESS = 'app/EditQuote/LOAD_XRM_DATA_SUCCESS';
+
+export const DEFAULT_ACTION = 'app/FavouriteHeader/DEFAULT_ACTION';
